# transaction-service/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Banking Transaction Service API",
    "version": "1.0.0",
    "description": "API documentation for the Transaction Microservice (includes customer, account, and transaction endpoints)"
  },
  "paths": {
    "/transactions/deposit": {
      "post": {
        "summary": "Deposit money into an account",
        "tags": [
          "Transactions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "account_id": {
                    "type": "integer"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "counterparty": {
                    "type": "string"
                  }
                },
                "required": [
                  "account_id",
                  "amount"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deposit successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or business error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/transactions/withdraw": {
      "post": {
        "summary": "Withdraw money from an account",
        "tags": [
          "Transactions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "account_id": {
                    "type": "integer"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "counterparty": {
                    "type": "string"
                  }
                },
                "required": [
                  "account_id",
                  "amount"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Withdrawal successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or business error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/transactions/statement/{account_id}": {
      "get": {
        "summary": "Get account statement (latest transactions)",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of transactions to return"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer"
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Statement retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or business error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}